Major update, still not quite there...
Wow, this was a big one. This work is to take the old approach, which
used minimal extra information along with only some data, and make it so
all forms carry with them additional information that can be used like a
type, or used to get line numbers for debugging.